+++ core/network/src/main/java/com/into/websoso/core/network/datasource/account/AccountApi.kt
@@ -1,9 +1,9 @@
 package com.into.websoso.core.network.datasource.account
 
-import com.into.websoso.core.network.datasource.account.model.KakaoLoginResponseDto
-import com.into.websoso.core.network.datasource.account.model.KakaoLogoutRequestDto
-import com.into.websoso.core.network.datasource.account.model.TokenReissueRequestDto
-import com.into.websoso.core.network.datasource.account.model.TokenReissueResponseDto
+import com.into.websoso.core.network.datasource.account.model.response.KakaoLoginResponseDto
+import com.into.websoso.core.network.datasource.account.model.request.KakaoLogoutRequestDto
+import com.into.websoso.core.network.datasource.account.model.request.TokenReissueRequestDto
+import com.into.websoso.core.network.datasource.account.model.response.TokenReissueResponseDto
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
