+++ core/network/src/main/java/com/into/websoso/core/network/datasource/account/DefaultAccountDataSource.kt
@@ -2,8 +2,8 @@ package com.into.websoso.core.network.datasource.account
 
 import com.into.websoso.core.auth.AuthPlatform
 import com.into.websoso.core.auth.AuthToken
-import com.into.websoso.core.network.datasource.account.model.KakaoLogoutRequestDto
-import com.into.websoso.core.network.datasource.account.model.TokenReissueRequestDto
+import com.into.websoso.core.network.datasource.account.model.request.KakaoLogoutRequestDto
+import com.into.websoso.core.network.datasource.account.model.request.TokenReissueRequestDto
 import com.into.websoso.data.account.datasource.AccountRemoteDataSource
 import com.into.websoso.data.account.model.AccountEntity
 import com.into.websoso.data.account.model.TokenEntity
