+++ core/network/src/main/java/com/into/websoso/core/network/datasource/account/model/response/KakaoLoginResponseDto.kt
@@ -1,4 +1,4 @@
-package com.into.websoso.core.network.datasource.account.model
+package com.into.websoso.core.network.datasource.account.model.response
 
 import com.into.websoso.data.account.model.AccountEntity
 import com.into.websoso.data.account.model.TokenEntity
