+++ core/network/src/main/java/com/into/websoso/core/network/datasource/account/model/response/TokenReissueResponseDto.kt
@@ -1,4 +1,4 @@
-package com.into.websoso.core.network.datasource.account.model
+package com.into.websoso.core.network.datasource.account.model.response
 
 import com.into.websoso.data.account.model.TokenEntity
 import kotlinx.serialization.SerialName
